+++ common/src/config/GameConfig.cpp
@@ -200,9 +200,8 @@ template<>
 bool is_valid_enum_value<GameConfig::ClampMode>(int value)
 {
     return value == GameConfig::AUTOMATIC
-        || value == GameConfig::FULL
-        || value == GameConfig::LEGACYPC
-        || value == GameConfig::LEGACYPS2;
+        || value == GameConfig::CLASSIC
+        || value == GameConfig::FULL;
 }
 
 template<>