+++ game_patch/misc/g_solid.cpp
@@ -232,39 +232,18 @@ CodeInjection level_load_lightmaps_color_conv_patch{
         rf::gr::LockInfo lock;
         if (!rf::gr::lock(lightmap->bm_handle, 0, &lock, rf::gr::LOCK_WRITE_ONLY))
             return;
-        // Apply legacy rf ps2 lightmap clamping if player has forced it
-        if (g_game_config.clamp_mode == GameConfig::LEGACYPS2)
-            {
-            xlog::debug("Applying Legacy RF (PS2) lightmap clamping");
-            for (int i = 0; i < lightmap->w * lightmap->h * 3; i += 3) {
-                uint8_t& r = lightmap->buf[i];
-                uint8_t& g = lightmap->buf[i + 1];
-                uint8_t& b = lightmap->buf[i + 2];
-                float sum = r + g + b;
-                if (0.57f < r / sum) {
-                    r = static_cast<uint8_t>(r * 0.75f);
-                    g = static_cast<uint8_t>(g * 0.88f);
-                    b = static_cast<uint8_t>(b * 0.88f);
-                }
-                if (0.57f < g / sum) {
-                    g = static_cast<uint8_t>(g * 0.87f);
-                }
-                if (0.57f < b / sum) {
-                    b = static_cast<uint8_t>(b * 0.87f);
-                }
-                if (190.0f < (r + g + b)) {
-                    r = static_cast<uint8_t>(r * 0.9f);
-                    g = static_cast<uint8_t>(g * 0.9f);
-                    b = static_cast<uint8_t>(b * 0.9f);
-                }
-            }
+        // apply clamp to Volition maps. Most Volition rfls ver = 180, community-made rfl all = 200
+        if (g_game_config.clamp_mode == GameConfig::AUTOMATIC && rf::level.version < 200) {
+            xlog::debug("Applying lightmap clamping to official Volition map");
+            for (int i = 0; i < lightmap->w * lightmap->h * 3; ++i)
+                lightmap->buf[i] = std::max(lightmap->buf[i], (uint8_t)(4 << 3)); // 32
         }
-        // Apply legacy rf pc lightmap clamping only if either player has forced it, or if player has selected automatic and the map was made prior to Sept 8, 2024 @ 00:00:00 UTC
-        else if (g_game_config.clamp_mode == GameConfig::LEGACYPC || (rf::level.level_timestamp < 1725753600 && g_game_config.clamp_mode == GameConfig::AUTOMATIC))
-            {
-            xlog::debug("Applying Legacy RF (PC) lightmap clamping");
-        for (int i = 0; i < lightmap->w * lightmap->h * 3; ++i)
-            lightmap->buf[i] = std::max(lightmap->buf[i], (uint8_t)(4 << 3)); // 32
+
+        // apply clamp to maps made before Sept 8, 2024 @ 00:00:00 UTC
+        else if (g_game_config.clamp_mode == GameConfig::CLASSIC && rf::level.level_timestamp < 1725753600) {
+            xlog::debug("Applying lightmap clamping to map made before 8/9/24");
+            for (int i = 0; i < lightmap->w * lightmap->h * 3; ++i)
+                lightmap->buf[i] = std::max(lightmap->buf[i], (uint8_t)(4 << 3)); // 32
         }
 
         bool success = bm_convert_format(lock.data, lock.format, lightmap->buf,
@@ -274,6 +253,31 @@ CodeInjection level_load_lightmaps_color_conv_patch{
 
         rf::gr::unlock(&lock);
     },
+};
+
+ConsoleCommand2 lighting_color_range_cmd{
+    "lighting_color_range", [] (std::optional<std::string> range_arg)
+    {
+        GameConfig::ClampMode new_mode;
+        if (range_arg == "automatic" || range_arg == "auto" || range_arg == "0") {
+            new_mode = GameConfig::ClampMode::AUTOMATIC;
+        }
+        else if (range_arg == "classic" || range_arg == "1") {
+            new_mode = GameConfig::ClampMode::CLASSIC;
+        }
+        else if (range_arg == "full" || range_arg == "2") {
+            new_mode = GameConfig::ClampMode::FULL;
+        }
+        else {
+            rf::console::printf("Invalid option. Valid options are auto, classic, full");
+            return;
+        }
+        g_game_config.clamp_mode = new_mode;
+        g_game_config.save();
+        rf::console::printf("Lighting color range set to %s", range_arg.value().c_str());
+    },
+    "Set lighting color range. Valid options are auto, classic, full. Only affects levels loaded after usage of this command.",
+    "lighting_color_range [option]",
 };
 
 CodeInjection shadow_render_one_injection{
@@ -499,4 +503,5 @@ void g_solid_do_patch()
     // Commands
     max_decals_cmd.register_cmd();
     dbg_room_clip_wnd_cmd.register_cmd();
+    lighting_color_range_cmd.register_cmd();
 }